Create Tomcatalyzer
Note that this script is looking for the string 'ms' in the log file and assumes that it is followed by the response time in milliseconds. This may not be the case in your log files and you may have to adjust the regular expression used to match the response time in the logs. Also, you need to replace the path of log files with the actual path where your files are located.

Also, This script will only provide the information you're looking for if your log files contain the necessary data. If the log files don't include response times or error messages, the script will not work as expected.